@@ -1,39 +1,61 @@
 {
-  "name": "image-manager-app",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "tsc -b && vite build",
-    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
-    "preview": "vite preview"
-  },
-  "dependencies": {
-    "@fontsource/montserrat": "^5.0.18",
-    "@reduxjs/toolkit": "^2.2.6",
-    "formik": "^2.4.6",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "react-icons": "^5.2.1",
-    "react-redux": "^9.1.2",
-    "yup": "^1.4.0"
-  },
-  "devDependencies": {
-    "@types/react": "^18.3.3",
-    "@types/react-dom": "^18.3.0",
-    "@typescript-eslint/eslint-plugin": "^7.17.0",
-    "@typescript-eslint/parser": "^7.17.0",
-    "@vitejs/plugin-react": "^4.3.1",
-    "autoprefixer": "^10.4.19",
-    "eslint": "^8.57.0",
-    "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-react": "^7.35.0",
-    "eslint-plugin-react-hooks": "^4.6.2",
-    "eslint-plugin-react-refresh": "^0.4.7",
-    "postcss": "^8.4.39",
-    "tailwindcss": "^3.4.6",
-    "typescript": "^5.2.2",
-    "vite": "^5.3.4"
-  }
+   "name": "image-manager-app",
+   "private": true,
+   "version": "0.0.0",
+   "type": "module",
+   "scripts": {
+      "dev": "vite",
+      "build": "tsc -b && vite build",
+      "preview": "vite preview",
+      "prepare": "husky install",
+      "lint": "eslint .",
+      "lint:fix": "eslint --fix .",
+      "prettier": "prettier --write .",
+      "prettier:check": "prettier --check ."
+   },
+   "husky": {
+      "hooks": {
+         "pre-commit": "lint-staged"
+      }
+   },
+   "lint-staged": {
+      "src/**/*.{js,jsx,ts,tsx}": [
+         "npm run lint"
+      ]
+   },
+   "dependencies": {
+      "@fontsource/montserrat": "^5.0.18",
+      "@reduxjs/toolkit": "^2.2.6",
+      "formik": "^2.4.6",
+      "react": "^18.3.1",
+      "react-dom": "^18.3.1",
+      "react-icons": "^5.2.1",
+      "react-redux": "^9.1.2",
+      "yup": "^1.4.0"
+   },
+   "devDependencies": {
+      "@eslint/js": "^9.7.0",
+      "@types/react": "^18.3.3",
+      "@types/react-dom": "^18.3.0",
+      "@typescript-eslint/eslint-plugin": "^7.17.0",
+      "@typescript-eslint/parser": "^7.17.0",
+      "@vitejs/plugin-react": "^4.3.1",
+      "autoprefixer": "^10.4.19",
+      "eslint": "^9.7.0",
+      "eslint-config-prettier": "^9.1.0",
+      "eslint-plugin-import": "^2.29.1",
+      "eslint-plugin-prettier": "^5.2.1",
+      "eslint-plugin-react": "^7.35.0",
+      "eslint-plugin-react-hooks": "^4.6.2",
+      "eslint-plugin-react-refresh": "^0.4.7",
+      "globals": "^15.8.0",
+      "husky": "^8.0.0",
+      "lint-staged": "^15.2.7",
+      "postcss": "^8.4.39",
+      "prettier": "^3.3.3",
+      "tailwindcss": "^3.4.6",
+      "typescript": "^5.2.2",
+      "typescript-eslint": "^7.17.0",
+      "vite": "^5.3.4"
+   }
 }
